+++ integrations/gitlab/CHANGELOG.md
@@ -12,7 +12,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Improvements
 
-- Bumped ocean version to ^0.4.0 (#1)
+- Bumped ocean version to ^0.4.0 and handle ONCE event listener (#1)
 
 
 0.1.23 (2023-10-30)